# interpreter/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Open Interpreter Server API",
    "description": "OpenAI-compatible API server for Open Interpreter with automatic OpenAPI documentation",
    "version": "0.4.3"
  },
  "paths": {
...
    "/v1/chat/completions": {
      "post": {
        "tags": [
          "default"
        ],
        "summary": "Create chat completion",
        "description": "OpenAI-compatible chat completion endpoint. Supports both streaming and non-streaming responses.",
        "operationId": "openai_chat_completions_v1_chat_completions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ValidationErrorModel"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatMessage": {
        "title": "ChatMessage",
        "required": [
          "role",
          "content"
        ],
        "type": "object",
        "properties": {
          "role": {
            "title": "Role",
            "type": "string",
            "description": "\u6d88\u606f\u89d2\u8272: user, assistant, system",
            "example": "user"
          },
          "content": {
            "title": "Content",
            "type": "string",
            "description": "\u6d88\u606f\u5185\u5bb9",
            "example": "Hello, how are you?"
          },
          "type": {
            "title": "Type",
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "\u6d88\u606f\u7c7b\u578b",
            "default": "message",
            "example": "message"
          }
        },
        "description": "\u804a\u5929\u6d88\u606f"
      },
...
      "ChatCompletionRequest": {
        "title": "ChatCompletionRequest",
        "required": [
          "messages"
        ],
        "type": "object",
        "properties": {
          "messages": {
            "title": "Messages",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "description": "\u6d88\u606f\u5217\u8868",
            "example": [
              {
                "content": "Hello, can you help me?",
                "role": "user"
              }
            ]
          },
          "model": {
            "title": "Model",
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "\u6a21\u578b\u540d\u79f0 (\u53ef\u9009\uff0c\u9ed8\u8ba4\u4f7f\u7528\u670d\u52a1\u5668\u914d\u7f6e)",
            "default": null,
            "example": "bedrock/anthropic.claude-3-5-sonnet-20240620-v1:0"
          },
          "stream": {
            "title": "Stream",
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "\u662f\u5426\u6d41\u5f0f\u54cd\u5e94",
            "default": false,
            "example": false
          },
          "session_id": {
            "title": "Session Id",
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "\u4f1a\u8bddID (\u53ef\u9009\uff0c\u7528\u4e8e\u4fdd\u6301\u5bf9\u8bdd\u4e0a\u4e0b\u6587)",
            "default": null,
            "example": "session-123"
          }
        },
        "description": "\u804a\u5929\u5b8c\u6210\u8bf7\u6c42",
        "example": {
          "messages": [
            {
              "content": "Hello, can you help me write a Python function?",
              "role": "user"
            }
          ],
          "model": "bedrock/anthropic.claude-3-5-sonnet-20240620-v1:0",
          "session_id": null,
          "stream": false
        }
      },
...
      "ValidationErrorModel": {
        "title": "ValidationErrorModel",
        "required": [
          "type",
          "loc",
          "msg",
          "input"
        ],
        "type": "object",
        "properties": {
          "type": {
            "title": "Error Type",
            "type": "string",
            "description": "A computer-readable identifier of the error type."
          },
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {},
            "description": "The error's location as a list."
          },
          "msg": {
            "title": "Message",
            "type": "string",
            "description": "A human readable explanation of the error."
          },
          "input": {
            "title": "Input",
            "description": "The input provided for validation."
          },
          "url": {
            "title": "URL",
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The URL to further information about the error.",
            "default": null
          },
          "ctx": {
            "title": "Error context",
            "anyOf": [
              {
                "type": "object",
                "additionalProperties": true
              },
              {
                "type": "null"
              }
            ],
            "description": "An optional object which contains values required to render the error message.",
            "default": null
          }
        }
      }
    },
    "securitySchemes": null
  }
}